Trying to find error in large lattice code.

+++ GluonicLQCD/config/configloader.cpp
@@ -66,7 +66,7 @@ namespace ConfigLoader {
 
         // Sub dimension setting
         if (!j["subDims"].empty()) {
-            std::vector<unsigned long int> tempN = {j["subDims"][0],j["subDims"][1],j["subDims"][2],j["subDims"][3]};
+            std::vector<unsigned int> tempN = {j["subDims"][0],j["subDims"][1],j["subDims"][2],j["subDims"][3]};
             Parameters::setN(tempN);
             Parameters::setSubLatticePreset(true);
             Parallel::Communicator::setN(tempN);
