+++ GluonicLQCD/io/fieldio.h
@@ -15,15 +15,15 @@ namespace IO {
         static const unsigned long int m_SU3Doubles;
         static const unsigned long int m_SU3Size;
 
-        static std::vector<unsigned long int> m_N;
+        static std::vector<unsigned int> m_N;
         static inline double reverseDouble(const double inDouble);
         static inline bool check_file_existence (const std::string fname);
     public:
         FieldIO();
         ~FieldIO();
         static void init();
-        static void writeFieldToFile(Lattice<SU3> *lattice, int configNumber);
-        static void writeDoublesFieldToFile(Lattice<double> lattice, int configNumber, std::string observable);
+        static void writeFieldToFile(Lattice<SU3> *lattice, unsigned long int configNumber);
+        static void writeDoublesFieldToFile(Lattice<double> lattice, unsigned long int configNumber, std::string observable);
         static void loadFieldConfiguration(std::string filename, Lattice<SU3> *lattice);
         static void loadChromaFieldConfiguration(std::string filename, Lattice<SU3> *lattice);
     };
